requirements.txt
# API y Servidor
fastapi==0.103.1
starlette==0.27.0  # Versión compatible con fastapi 0.103.1
uvicorn==0.23.2
pydantic==2.10.0

# OpenAI e IA
openai>=1.81.0
langchain==0.0.312
anyio==3.7.1

# Procesamiento de voz
elevenlabs
librosa==0.10.1
soundfile==0.12.1

# Base de datos y caché
sqlalchemy==2.0.22
supabase
asyncpg==0.28.0

# Utilidades
python-dotenv==1.0.0
httpx==0.24.1  # Versión compatible con starlette 0.27.0
pyjwt==2.8.0
tenacity==8.2.3

# Observabilidad
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
opentelemetry-instrumentation-asyncpg==0.44b0
opentelemetry-instrumentation-logging==0.44b0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1 